+++ src/org/nschmidt/abalone/PlayerLogic.java
@@ -1,10 +1,9 @@
 package org.nschmidt.abalone;
 
 import java.awt.Color;
-import java.math.BigInteger;
 
 public interface PlayerLogic {
     public Player switchPlayer();
-    public BigInteger getNumber();
+    public int getNumber();
     public Color getColor();
 }
